Move Telugu char generator to fixture folder
